fix: codex+agy multi-tool whole-codebase audit — 145 confirmed bugs, 4-agent-reviewed
Independent codex exec + agy multi-tool review of every module in xpcsjax (29 file-bounded units, ~85K lines), self-verified against real code, narrowed 168 raw findings to 145 confirmed bugs, fixed module by module.

Then ran a 4-agent review (code-reviewer, pr-test-analyzer, silent-failure-hunter, type-design-analyzer) against the fix diff itself, which found 3 blocking issues and ~13 regressions the fixes had introduced — all addressed (2 reverted where no working verification gate existed, the rest fixed with empirical repros and 8 new regression tests).

Reconciled with main's concurrent PR #18 (which bundled several overlapping bug fixes despite its "docs" description) — 5 real merge conflicts, each resolved by combining both sides rather than discarding either; caught and fixed one additional regression (`safe_bool(None, ...)`) that the reconciliation itself introduced.

make verify: 2466 passed, lint clean, mypy advisory clean.

+++ xpcsjax/viz/diagnostics.py
@@ -127,7 +127,10 @@ def compute_diagonal_overlay_stats(
     # not inf, so an inf entry would otherwise poison the mean.
     pair_mask = np.isfinite(raw_diag) & np.isfinite(fitted_diag)
     if pair_mask.any():
-        diff = fitted_diag[pair_mask] - raw_diag[pair_mask]
+        # float64 promotion: integer-typed c2 inputs would otherwise square
+        # in their own (possibly overflowing) integer dtype before np.mean
+        # gets a chance to upcast.
+        diff = fitted_diag[pair_mask].astype(np.float64) - raw_diag[pair_mask].astype(np.float64)
         fitted_rmse = float(np.sqrt(np.mean(diff**2)))
     else:
         fitted_rmse = float("nan")
